+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/configuration/ValidatorKey.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2018 The Exonum Team
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.exonum.binding.common.configuration;
+
+import com.exonum.binding.common.hash.HashCode;
+import com.google.auto.value.AutoValue;
+import com.google.gson.annotations.SerializedName;
+import com.ryanharter.auto.value.gson.GenerateTypeAdapter;
+
+@AutoValue
+@GenerateTypeAdapter
+public abstract class ValidatorKey {
+  @SerializedName("consensus_key")
+  public abstract HashCode consensusKey();
+
+  @SerializedName("service_key")
+  public abstract HashCode serviceKey();
+
+  public static ValidatorKey create(HashCode consensusKey, HashCode serviceKey) {
+    return new AutoValue_ValidatorKey(consensusKey, serviceKey);
+  }
+}