+++ exonum-java-binding-core/src/test/java/com/exonum/binding/blockchain/BlockchainTest.java
@@ -22,8 +22,13 @@ import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import com.exonum.binding.common.configuration.ConsensusConfig;
+import com.exonum.binding.common.configuration.StoredConfiguration;
+import com.exonum.binding.common.configuration.ValidatorKey;
+import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.storage.indices.ListIndexProxy;
 import com.exonum.binding.storage.indices.ProofListIndexProxy;
+import java.util.Collections;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -66,4 +71,30 @@ class BlockchainTest {
     assertThat(blockchain.getBlockTransactions(1L)).isEqualTo(mockListIndex);
   }
 
+  @Test
+  void getActualConfiguration() {
+    StoredConfiguration configuration = createConfiguration();
+    when(mockSchema.getActualConfiguration()).thenReturn(configuration);
+
+    assertThat(blockchain.getActualConfiguration()).isEqualTo(configuration);
+  }
+
+  private StoredConfiguration createConfiguration() {
+    return StoredConfiguration.create(
+        HashCode.fromString("11"),
+        1,
+        Collections.singletonList(
+            ValidatorKey.create(HashCode.fromString("22"), HashCode.fromString("33"))
+        ),
+        ConsensusConfig.create(
+            1,
+            2,
+            3,
+            4,
+            5,
+            6,
+            7,
+            8)
+    );
+  }
 }